src/square-checkout-api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
squareup